feat: Add global object.
