+++ casdoorsdk/email.go
@@ -1,4 +1,4 @@
-// Copyright 2021 The Casdoor Authors. All Rights Reserved.
+// Copyright 2023 The Casdoor Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -48,8 +48,4 @@ func (c *Client) SendEmail(title string, content string, sender string, receiver
 	}
 
 	return nil
-}
-
-func SendEmail(title string, content string, sender string, receivers ...string) error {
-	return globalClient.SendEmail(title, content, sender, receivers...)
 }
