+++ casdoorsdk/role.go
@@ -1,4 +1,4 @@
-// Copyright 2022 The Casdoor Authors. All Rights Reserved.
+// Copyright 2023 The Casdoor Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -55,10 +55,6 @@ func (c *Client) GetRoles() ([]*Role, error) {
 	return roles, nil
 }
 
-func GetRoles() ([]*Role, error) {
-	return globalClient.GetRoles()
-}
-
 func (c *Client) GetPaginationRoles(p int, pageSize int, queryMap map[string]string) ([]*Role, int, error) {
 	queryMap["owner"] = c.OrganizationName
 	queryMap["p"] = strconv.Itoa(p)
@@ -76,10 +72,6 @@ func (c *Client) GetPaginationRoles(p int, pageSize int, queryMap map[string]str
 		return nil, 0, errors.New("response data format is incorrect")
 	}
 	return roles, int(response.Data2.(float64)), nil
-}
-
-func GetPaginationRoles(p int, pageSize int, queryMap map[string]string) ([]*Role, int, error) {
-	return globalClient.GetPaginationRoles(p, pageSize, queryMap)
 }
 
 func (c *Client) GetRole(name string) (*Role, error) {
@@ -102,17 +94,9 @@ func (c *Client) GetRole(name string) (*Role, error) {
 	return role, nil
 }
 
-func GetRole(name string) (*Role, error) {
-	return globalClient.GetRole(name)
-}
-
 func (c *Client) UpdateRole(role *Role) (bool, error) {
 	_, affected, err := c.modifyRole("update-role", role, nil)
 	return affected, err
-}
-
-func UpdateRole(role *Role) (bool, error) {
-	return globalClient.UpdateRole(role)
 }
 
 func (c *Client) UpdateRoleForColumns(role *Role, columns []string) (bool, error) {
@@ -120,24 +104,12 @@ func (c *Client) UpdateRoleForColumns(role *Role, columns []string) (bool, error
 	return affected, err
 }
 
-func UpdateRoleForColumns(role *Role, columns []string) (bool, error) {
-	return globalClient.UpdateRoleForColumns(role, columns)
-}
-
 func (c *Client) AddRole(role *Role) (bool, error) {
 	_, affected, err := c.modifyRole("add-role", role, nil)
 	return affected, err
 }
 
-func AddRole(role *Role) (bool, error) {
-	return globalClient.AddRole(role)
-}
-
 func (c *Client) DeleteRole(role *Role) (bool, error) {
 	_, affected, err := c.modifyRole("delete-role", role, nil)
 	return affected, err
-}
-
-func DeleteRole(role *Role) (bool, error) {
-	return globalClient.DeleteRole(role)
 }
