+++ casdoorsdk/role_global.go
@@ -1,0 +1,43 @@
+// Copyright 2023 The Casdoor Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package casdoorsdk
+
+func GetRoles() ([]*Role, error) {
+	return globalClient.GetRoles()
+}
+
+func GetPaginationRoles(p int, pageSize int, queryMap map[string]string) ([]*Role, int, error) {
+	return globalClient.GetPaginationRoles(p, pageSize, queryMap)
+}
+
+func GetRole(name string) (*Role, error) {
+	return globalClient.GetRole(name)
+}
+
+func UpdateRole(role *Role) (bool, error) {
+	return globalClient.UpdateRole(role)
+}
+
+func UpdateRoleForColumns(role *Role, columns []string) (bool, error) {
+	return globalClient.UpdateRoleForColumns(role, columns)
+}
+
+func AddRole(role *Role) (bool, error) {
+	return globalClient.AddRole(role)
+}
+
+func DeleteRole(role *Role) (bool, error) {
+	return globalClient.DeleteRole(role)
+}
