+++ casdoorsdk/url.go
@@ -1,4 +1,4 @@
-// Copyright 2021 The Casdoor Authors. All Rights Reserved.
+// Copyright 2023 The Casdoor Authors. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -29,10 +29,6 @@ func (c *Client) GetSignupUrl(enablePassword bool, redirectUri string) string {
 	}
 }
 
-func GetSignupUrl(enablePassword bool, redirectUri string) string {
-	return globalClient.GetSignupUrl(enablePassword, redirectUri)
-}
-
 func (c *Client) GetSigninUrl(redirectUri string) string {
 	// origin := "https://door.casbin.com"
 	// redirectUri := fmt.Sprintf("%s/callback", origin)
@@ -40,10 +36,6 @@ func (c *Client) GetSigninUrl(redirectUri string) string {
 	state := c.ApplicationName
 	return fmt.Sprintf("%s/login/oauth/authorize?client_id=%s&response_type=code&redirect_uri=%s&scope=%s&state=%s",
 		c.Endpoint, c.ClientId, url.QueryEscape(redirectUri), scope, state)
-}
-
-func GetSigninUrl(redirectUri string) string {
-	return globalClient.GetSigninUrl(redirectUri)
 }
 
 func (c *Client) GetUserProfileUrl(userName string, accessToken string) string {
@@ -54,18 +46,10 @@ func (c *Client) GetUserProfileUrl(userName string, accessToken string) string {
 	return fmt.Sprintf("%s/users/%s/%s%s", c.Endpoint, c.OrganizationName, userName, param)
 }
 
-func GetUserProfileUrl(userName string, accessToken string) string {
-	return globalClient.GetUserProfileUrl(userName, accessToken)
-}
-
 func (c *Client) GetMyProfileUrl(accessToken string) string {
 	param := ""
 	if accessToken != "" {
 		param = fmt.Sprintf("?access_token=%s", accessToken)
 	}
 	return fmt.Sprintf("%s/account%s", c.Endpoint, param)
-}
-
-func GetMyProfileUrl(accessToken string) string {
-	return globalClient.GetMyProfileUrl(accessToken)
 }
